backend-flask/requirements.txt
flask
flask-cors

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-distro

aws-xray-sdk

watchtower

blinker
rollbar
